% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.nn.distance.functions.R
\name{dist.fun.bubble}
\alias{dist.fun.bubble}
\title{Bubble distance functions for topological k-NN classifier}
\usage{
dist.fun.bubble(x, sigma = 1.1)
}
\arguments{
\item{x}{Distance or \code{numeric} vector or matrix of distances.}

\item{sigma}{Maximum distance to be considered. Default is 1.1.}
}
\value{
Distance-dependent weight.
}
\description{
The function is used as distance-dependent weight \eqn{w} for k-NN voting.
}
\details{
The function returns 1.0 for \eqn{0 < x \le \sigma} and 0.0 for \eqn{x > \sigma}.
}

