% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.nn.multitrain.R
\name{enrich.dirty}
\alias{enrich.dirty}
\title{enrich training set with dirty mapped samples}
\usage{
enrich.dirty(x, model, multiple)
}
\arguments{
\item{x}{training data}

\item{model}{SOMnn model}

\item{multiple}{enhancement factor for dirty samples}
}
\description{
Maps x to the SOM defined in model and makes a list of dirty neurons
(i.e. neurons with more then one class label mapped).
All training samples in these neurons are added to the training set
to enhance their training.
}
\keyword{internal}
