% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.nn.multitrain.R
\name{get.border.neurons}
\alias{get.border.neurons}
\title{Get border neurons.}
\usage{
get.border.neurons(p, classes, model, distance = 1.1)
}
\arguments{
\item{p}{prediction for training data set}

\item{classes}{vector of true class lables for prediction}

\item{model}{Object of class type \code{SOMnn}}

\item{distance}{maximum distance of 2 neurons to be the border. Default 1.1:
only direct neighbours.}
}
\value{
numeric vector with the indices of all border neurons.
}
\description{
Returns a list of neurons which are on the border between 2 or more classes.
}
\details{
The function analyses all pairs of neurons with distance <= \code{distance}.
If samples represented by the pair belong to more than one class, both neurons
are added to the list.
}
\keyword{internal}
