% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SOMnn.R
\name{hexbinpie}
\alias{hexbinpie}
\title{Plots the hexagonals and pi charts.
Adapted code from package somplot.}
\usage{
hexbinpie(
  x,
  y,
  kat,
  xbnds = range(x),
  ybnds = range(y),
  hbc = NA,
  pal = NA,
  hex = "gray",
  circ = "gray50",
  cnt = "black",
  show.counter.border,
  ...
)
}
\description{
Plots the hexagonals and pi charts.
Adapted code from package somplot.
}
\keyword{internal}
