% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.nn.util.R
\name{make.codes.grid}
\alias{make.codes.grid}
\title{Makes a data.frame with codes coordinates}
\usage{
make.codes.grid(xdim, ydim, topo = "hexagonal")
}
\description{
Coordinates of neurons of a som are calculated by 
calling \code{\link[class]{somgrid}} to be consistent with
other som/kohonen packages.
}
\keyword{internal}

