% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SOMnn.R
\name{plot.predictions}
\alias{plot.predictions}
\title{Plots predicted samples as points into a plotted som.}
\usage{
plot.predictions(grid, predict, pch.col, pch, ...)
}
\description{
Plots predicted samples as points into a plotted som.
}
\keyword{internal}

