% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.nn.util.R
\name{round.probabilities}
\alias{round.probabilities}
\title{Advanced rounding of vectors}
\usage{
\method{round}{probabilities}(x, digits = 2)
}
\arguments{
\item{x}{\code{numeric} vector of values.}

\item{digits}{demanded precision}
}
\description{
Rounds a vector of probabilities preserving their sum.
}
\details{
In general, if a vector of floating point values is rounded,
the sum is not preserverd.
For a vector of probabilities (which sum up to 1.0), this may lead to
strange results.
This function rounds all values of the vector and takes care, that
the sum ist not changed (with a precision given in \code{digits}).
}
