% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.nn.export.kohonen.R
\name{som.nn.export.kohonen}
\alias{som.nn.export.kohonen}
\title{Export a som.nn model as object of type \code{kohonen}}
\usage{
som.nn.export.kohonen(model, train)
}
\arguments{
\item{model}{model of type \code{SOMnn}.}

\item{train}{training data}
}
\value{
list of type \code{kohonen} with the trained som.
                See \code{\link[kohonen]{som}} for details.
}
\description{
An existing model of type \code{SOMnn} is exported as
object of type \code{kohonen} for use with the tools of the 
package \code{kohonen}.
}
\details{
Training data is necessary to generate the kohonen object.
}

