% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.nn.export.som.R
\name{som.nn.export.som}
\alias{som.nn.export.som}
\title{Export a som.nn model as object of type \code{SOM}}
\usage{
som.nn.export.som(model)
}
\arguments{
\item{model}{model of type \code{SOMnn}.}
}
\value{
List of type \code{SOM} with the trained som.
                See \code{\link[class]{SOM}} for details.
}
\description{
An existing model of type \code{SOMnn} is exported as
object of type \code{SOM} for use with the tools of the 
package \code{class}.
}

