% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.nn.util.R
\name{som.nn.round.votes}
\alias{som.nn.round.votes}
\title{Rounds a dataframe with vectors of votes for SOMnn}
\usage{
som.nn.round.votes(votes, classes, digits = 2)
}
\arguments{
\item{votes}{\code{data.frame} with rows of class probabilities.}

\item{classes}{\code{character} vector with name of categories.
Names must match the column names of probabilities to be
rounded.}

\item{digits}{precision; default = 2.}
}
\value{
\code{data.frame} with roundes rows of class probabilities.
        other columns are not affected.
}
\description{
Each row of the \code{data.frame} may sum up to 1.0
before and after rounding.
Rounding is performed with \code{round.probabilities}.
}
\keyword{internal}

