% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.nn.run.kernel.R
\name{som.nn.run.kernel}
\alias{som.nn.run.kernel}
\title{calls the specified kernel for som training.}
\usage{
som.nn.run.kernel(data, classes = "no classes", kernel = c("internal",
  "SOM"), xdim, ydim, rlen = 100, alpha = 0.05, radius = 1, init,
  toroidal = FALSE)
}
\arguments{
\item{data}{\code{numeric} matrix or data.frame with training data. 
Only numeric columns of data.frame are used for training.}

\item{classes}{\code{character} vector with class labels (only necessary for
supervised training kernels).}

\item{kernel}{kernel to be used}

\item{xdim}{number of neurons in x}

\item{ydim}{number of neurons in y}

\item{alpha}{initial learning rate (decreased to 0).}

\item{radius}{initial radius (decreased to 1).}

\item{init}{\code{numeric} matrix or data.frame with codes for initialisation.}

\item{toroidal}{true if doughnut-shaped som.}

\item{len}{number of steps to be trained (steps - not epochs!).}
}
\value{
list with elements \code{codes} and \code{grid}.
}
\description{
calls the specified kernel for som training.
}
\keyword{internal}

