% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.nn.som.experimental.R
\name{som.nn.som.experimental}
\alias{som.nn.som.experimental}
\title{Work hourse for som training.}
\usage{
som.nn.som.experimental(data, grid, rlen = 100, alpha = 0.05, radius, init,
  toroidal = FALSE)
}
\arguments{
\item{data}{matrix with training data.}

\item{grid}{somgrid object}

\item{alpha}{learning rate c(first, last).}

\item{radius}{radius c(first, last).}

\item{init}{codes for initialisation.}

\item{toroidal}{true if doughnut-shaped som.}

\item{len}{number of steps to be trained (steps - not epochs!).}
}
\value{
S3 object of type \code{kohonen} with the trained som.
}
\description{
Function is the kernel \code{internal} for som training, implemented
in pure R.
}
\keyword{internal}

