% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.nn.som.gaussian.R
\name{som.nn.som.gaussian}
\alias{som.nn.som.gaussian}
\title{Gaussian kernel for som training.}
\usage{
som.nn.som.gaussian(data, grid, len = 100, alpha = 0.05, radius, init,
  toroidal = FALSE)
}
\arguments{
\item{data}{matrix with training data.}

\item{grid}{somgrid object}

\item{len}{number of steps to be trained (steps - not epochs!).}

\item{alpha}{learning rate.}

\item{radius}{radius.}

\item{init}{codes for initialisation.}

\item{toroidal}{true if doughnut-shaped som.}
}
\value{
S3 object of type \code{kohonen} with the trained som.
}
\description{
Function is the kernel \code{gaussian} for som training, implemented
in pure R.
}
\keyword{internal}

