% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.nn.visual.R
\name{som.nn.visual.one}
\alias{som.nn.visual.one}
\title{Maps one vector to the SOM}
\usage{
som.nn.visual.one(one, codes)
}
\arguments{
\item{one}{\code{numeric} vector to be mapped}

\item{codes}{\code{numeric matrix} of codebook vectors with
one code per row}
}
\value{
vector with 2 elements: index of winner and qerror
}
\description{
Working hourse function for visual.
}
\keyword{internal}
