% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.nn.util.R
\name{som.nn.max.row}
\alias{som.nn.max.row}
\title{Special version of maximum finder for SOMnn}
\usage{
som.nn.max.row(x)
}
\arguments{
\item{x}{data.frame or matrix}
}
\value{
index of max value for each row or 0, if more
          than one element has the same maximum value.
}
\description{
Returns the index of the column with the maximum value for each row of a
data.frame.
If more than one element has the same maximum value, 0 is returned.
}
\keyword{internal}

